requirements.txt
opencv-python
numpy
imutils
python-telegram-bot
psycopg2-binary==2.9.7